api-sql-testing/sql/requirements.txt
requests==2.31.0
mysql-connector-python==8.1.1
python-dotenv==1.1.1
pandas==2.2.1